Added Exception Handling and new JPA function in repo

+++ src/main/java/com/sumerge/springtask/service/CourseService.java
@@ -1,9 +1,12 @@
 package com.sumerge.springtask.service;
 
 import com.sumerge.springtask.DTOs.CourseDTO;
+import com.sumerge.springtask.exceptions.CourseAlreadyExistsException;
+import com.sumerge.springtask.exceptions.NoCoursesAvailableException;
 import com.sumerge.springtask.mappers.CourseMapper;
 import com.sumerge.springtask.model.Course;
 import com.sumerge.springtask.repository.CourseRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -40,25 +43,40 @@ public class CourseService {
 
     public CourseDTO findById(Long id) {
         Course course = courseRepository.findFirstById(id);
+        if (course == null) {
+            throw new EntityNotFoundException("Course with id: " + id + " not found");
+        }
         return getCourseDTO(course);
     }
 
     public Page<CourseDTO> findAll(int page, int size) {
         Pageable pageable = PageRequest.of(page, size);
         Page<Course> courses = courseRepository.findAll(pageable);
+        if (courses.isEmpty()) {
+            throw new NoCoursesAvailableException("No courses available");
+        }
         return courses.map(this::getCourseDTO);
     }
 
     public void save(CourseDTO courseDTO) {
+        if(courseRepository.existsByCourseName(courseDTO.getCourse_name())){
+            throw new CourseAlreadyExistsException("This course already exists");
+        }
         Course course = getCourse(courseDTO);
         courseRepository.save(course);
     }
 
     public void delete(Long id) {
+        if(!courseRepository.existsById(id)){
+            throw new EntityNotFoundException("Course with id: " + id + " not found");
+        }
         courseRepository.deleteById(id);
     }
 
     public void update(Long id, CourseDTO courseDTO) {
+        if(!courseRepository.existsById(id)){
+            throw new EntityNotFoundException("Course with id: " + id + " not found");
+        }
         Course course = getCourse(courseDTO);
         Course toUpdate = courseRepository.findFirstById(id);
         toUpdate.setCourse_name(course.getCourse_name());
